requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp==2025.11.12
python-dotenv